+++ app/controllers/api/json/visualizations_controller.rb
@@ -461,51 +461,6 @@ class Api::Json::VisualizationsController < Api::ApplicationController
     ]
   end
 
-  def rows_and_sizes_for(table_data)
-    data = Hash.new
-
-    tables_data_per_schema = table_data.group_by { |td| td[:schema] }
-
-    tables_data_per_schema.each { |schema, tables_data| 
-      table_names = tables_data.map { |td| td[:name] }
-
-      begin
-        rows = current_user.in_database.fetch(%Q{
-          SELECT
-            relname AS table_name,
-            pg_total_relation_size('"' || ? || '"."' || relname || '"') AS total_relation_size,
-            reltuples::integer AS reltuples
-          FROM pg_class
-          WHERE relname in ?
-        },
-        schema,
-        table_names
-        ).all
-
-        rows.map { |row|
-          data[row[:table_name]] = {
-            size: row[:total_relation_size].to_i / 2,
-            rows: row[:reltuples]
-          }
-        }
-      rescue => e
-        # INFO: we don't want request to fail because of SQL error
-        CartoDB.notify_exception(e)
-      end
-    }
-
-    # Fill missing table data
-    # don't break whole dashboard
-    table_data.select { |td| data[td[:name]].nil? }.each { |table|
-        data[table[:name]] = {
-          size: nil,
-          rows: nil
-        }
-    }
-
-    data
-  end
-
   # This only allows to authenticate if sending an API request to username.api_key subdomain,
   # but doesn't breaks the request if can't authenticate
   def optional_api_authorization
@@ -516,17 +471,14 @@ class Api::Json::VisualizationsController < Api::ApplicationController
 
   def index_not_logged_in
     public_visualizations = []
+    total_liked_entries = 0
+    total_shared_entries = 0
     user = User.where(username: CartoDB.extract_subdomain(request)).first
 
     unless user.nil?
       filtered_params = params.dup.merge(scope_for(user))
-      filtered_params['exclude_shared'] = true
-      filtered_params['privacy'] = Visualization::Member::PRIVACY_PUBLIC
-      filtered_params.delete('locked')
-      filtered_params.delete('map_id')
-      collection = Visualization::Collection.new.fetch(
-        filtered_params
-      )
+      filtered_params[Visualization::Collection::FILTER_UNAUTHENTICATED] = true
+      collection = Visualization::Collection.new.fetch(filtered_params)
 
       public_visualizations  = collection.map { |vis|
         begin
@@ -539,11 +491,16 @@ class Api::Json::VisualizationsController < Api::ApplicationController
           puts exception.to_s + exception.backtrace.join("\n")
         end
       }.compact
+
+      total_liked_entries = collection.total_liked_entries
+      total_shared_entries = collection.total_shared_entries
     end
 
     response = {
       visualizations: public_visualizations,
-      total_entries:  public_visualizations.length
+      total_entries:  public_visualizations.length,
+      total_likes:    total_liked_entries,
+      total_shared:   total_shared_entries
     }
     render_jsonp(response)
   end
@@ -567,7 +524,6 @@ class Api::Json::VisualizationsController < Api::ApplicationController
       end
     }.compact
     synchronizations = synchronizations_by_table_name(table_data)
-    rows_and_sizes   = rows_and_sizes_for(table_data)
     representation  = collection.map { |vis|
       begin
         vis.to_hash(
@@ -575,17 +531,18 @@ class Api::Json::VisualizationsController < Api::ApplicationController
           table_data: !(params[:table_data] =~ /false/),
           user:       current_user,
           table:      vis.table,
-          synchronization: synchronizations[vis.name],
-          rows_and_sizes: rows_and_sizes
+          synchronization: synchronizations[vis.name]
         )
       rescue => exception
         puts exception.to_s + exception.backtrace.join("\n")
       end
     }.compact
 
-    response        = {
+    response = {
       visualizations: representation,
-      total_entries:  collection.total_entries
+      total_entries:  collection.total_entries,
+      total_likes:    collection.total_liked_entries,
+      total_shared:   collection.total_shared_entries
     }
     current_user.update_visualization_metrics
     render_jsonp(response)