+++ app/models/visualization/collection.rb
@@ -28,6 +28,13 @@ module CartoDB
       FILTER_SHARED_NO = 'no'
       FILTER_SHARED_ONLY = 'only'
 
+      FILTER_UNAUTHENTICATED = :unauthenticated
+      ORDERING_RELATED_ATTRIBUTES = [:likes, :mapviews, :row_count, :size]
+
+      # Same as services/data-repository/backend/sequel.rb
+      PAGE          = 1
+      PER_PAGE      = 300
+
       ALL_RECORDS = 999999
 
       def initialize(attributes={}, options={})
@@ -38,6 +45,11 @@ module CartoDB
           repository:   options.fetch(:repository, Visualization.repository),
           member_class: Member
         )
+        @can_paginate = true
+        @lazy_order_by = nil
+        @unauthenticated_flag = false
+        @user_id = nil
+        @type = nil
       end
 
       DataRepository::Collection::INTERFACE.each do |method_name|
@@ -54,6 +66,8 @@ module CartoDB
       # - only_shared forces to use different flow because if there are no shared there's nothing else to do
       # - locked filter has special behaviour
       def fetch(filters={})
+        @user_id = filters.fetch(:user_id, nil)
+        filters = restrict_filters_if_unauthenticated(filters)
         dataset = compute_sharing_filter_dataset(filters)
 
         if dataset.nil?
@@ -63,11 +77,23 @@ module CartoDB
           dataset = apply_filters(dataset, filters)
 
           @total_entries = dataset.count
-          dataset = repository.paginate(dataset, filters, @total_entries)
 
-          collection.storage = Set.new(dataset.map { |attributes|
-            Visualization::Member.new(attributes)
-          })
+          if @can_paginate
+            dataset = repository.paginate(dataset, filters, @total_entries)
+            collection.storage = Set.new(dataset.map { |attributes|
+              Visualization::Member.new(attributes)
+            })
+          else
+            items = dataset.map { |attributes|
+              Visualization::Member.new(attributes)
+            }
+            items = lazy_order_by(items, @lazy_order_by)
+            # Manual paging
+            page = (filters.delete(:page) || PAGE).to_i
+            per_page = (filters.delete(:per_page) || PER_PAGE).to_i
+            items = items.slice((page - 1) * per_page, per_page)
+            collection.storage = Set.new(items)
+          end
         end
 
         self
@@ -75,16 +101,16 @@ module CartoDB
 
       def delete_if(&block)
         collection.delete_if(&block)
+      end
+
+      # This method is not used for anything but called from the DataRepository::Collection interface above
+      def store
         self
       end
 
       def count_query(filters={})
         dataset = compute_sharing_filter_dataset(filters)
         dataset.nil? ? 0 : apply_filters(dataset, filters).count
-      end
-
-      def store
-        self
       end
 
       def destroy
@@ -97,11 +123,108 @@ module CartoDB
       end
 
       # Warning, this is a cached count, do not use if adding/removing collection items
+      # @throws KeyError
+      def total_shared_entries(raise_on_error = false)
+        user_shared_count = org_shared_count = 0
+
+        unless @unauthenticated_flag
+          if @user_id.nil? && raise_on_error
+            raise KeyError.new("Can't retrieve shared count without specifying user id")
+          else
+            user_shared_count = CartoDB::SharedEntity.select(:entity_id)
+              .where(:recipient_id => @user_id,
+                   :entity_type => CartoDB::SharedEntity::ENTITY_TYPE_VISUALIZATION,
+                   :recipient_type => CartoDB::SharedEntity::RECIPIENT_TYPE_USER)
+            if @type.nil?
+              user_shared_count = user_shared_count.join(:visualizations,
+                                                         :visualizations__id.cast(:uuid) => :entity_id)
+            else
+              user_shared_count = user_shared_count.join(:visualizations,
+                                                         :visualizations__id.cast(:uuid) => :entity_id,
+                                                         :type => @type)
+            end
+            user_shared_count = user_shared_count.count
+
+            user = User.where(id: @user_id).first
+            if user.nil? || user.organization.nil?
+              org_shared_count = 0
+            else
+              org_shared_count = CartoDB::SharedEntity.select(:entity_id)
+              .where(:recipient_id => user.organization_id,
+                     :entity_type => CartoDB::SharedEntity::ENTITY_TYPE_VISUALIZATION,
+                     :recipient_type => CartoDB::SharedEntity::RECIPIENT_TYPE_ORGANIZATION)
+              if @type.nil?
+                org_shared_count = org_shared_count.join(:visualizations,
+                                                         :visualizations__id.cast(:uuid) => :entity_id)
+              else
+                org_shared_count = org_shared_count.join(:visualizations,
+                                                         :visualizations__id.cast(:uuid) => :entity_id,
+                                                          :type => @type)
+              end
+              org_shared_count = org_shared_count.count
+            end
+          end
+        end
+        user_shared_count + org_shared_count
+      end
+
+      # @throws KeyError
+      def total_liked_entries(raise_on_error = false)
+        if @user_id.nil? && raise_on_error
+            raise KeyError.new("Can't retrieve likes count without specifying user id")
+        else
+          # Inner join with visualizations to filter by type and user, then applied a distinct count
+          if @unauthenticated_flag
+            likes_count = CartoDB::Like.select(:subject)
+            if @type.nil?
+              likes_count = likes_count.join(:visualizations,
+                                             :id.cast(:uuid) => :subject,
+                                             :user_id => @user_id,
+                                             :privacy => Visualization::Member::PRIVACY_PUBLIC)
+            else
+              likes_count = likes_count.join(:visualizations,
+                                             :id.cast(:uuid) => :subject,
+                                             :user_id => @user_id,
+                                             :type => @type,
+                                             :privacy => Visualization::Member::PRIVACY_PUBLIC)
+            end
+            likes_count = likes_count.distinct.count
+          else
+            likes_count = CartoDB::Like.select(:subject)
+            if @type.nil?
+              likes_count = likes_count.join(:visualizations,
+                                             :id.cast(:uuid) => :subject,
+                                             :user_id => @user_id)
+            else
+              likes_count = likes_count.join(:visualizations,
+                                             :id.cast(:uuid) => :subject,
+                                             :user_id => @user_id,
+                                             :type => @type)
+            end
+            likes_count = likes_count.distinct.count
+          end
+        end
+        likes_count
+      end
+
       attr_reader :total_entries
 
       private
 
       attr_reader :collection
+
+      # If special filter unauthenticated: true is present, will restrict data
+      def restrict_filters_if_unauthenticated(filters)
+        @unauthenticated_flag = false
+        unless filters.delete(FILTER_UNAUTHENTICATED).nil?
+          filters[:exclude_shared] = true
+          filters[:privacy] = Visualization::Member::PRIVACY_PUBLIC
+          filters.delete(:locked)
+          filters.delete(:map_id)
+          @unauthenticated_flag = true
+        end
+        filters
+      end
 
       def compute_sharing_filter_dataset(filters)
         shared_filter = filters.delete(:shared)
@@ -138,18 +261,67 @@ module CartoDB
       end
 
       def apply_filters(dataset, filters)
+        @type = filters.fetch(:type, nil)
         dataset = repository.apply_filters(dataset, filters, AVAILABLE_FILTERS)
         dataset = filter_by_tags(dataset, tags_from(filters))
         dataset = filter_by_partial_match(dataset, filters.delete(:q))
         dataset = filter_by_kind(dataset, filters.delete(:exclude_raster))
-        dataset = order(dataset, filters.delete(:o))
+        order(dataset, filters.delete(:order))
+      end
+
+      # Note: Not implemented ascending order for now, all are descending sorts
+      def lazy_order_by(objects, field)
+        case field
+          when :likes
+            objects.sort! { |obj_a, obj_b|
+              obj_b.likes.count <=> obj_a.likes.count
+            }
+          when :mapviews
+            objects.sort! { |obj_a, obj_b|
+              # Stats have format [ date, value ]
+              obj_b.stats.collect{|o| o[1] }.reduce(:+) <=> obj_a.stats.collect{|o| o[1] }.reduce(:+)
+            }
+          when :row_count
+            objects.sort! { |obj_a, obj_b|
+              a_rows = (obj_a.table.nil? ? 0 : obj_a.table.rows_and_size.fetch(:rows)) || 0
+              b_rows = (obj_b.table.nil? ? 0 : obj_b.table.rows_and_size.fetch(:rows)) || 0
+              b_rows <=> a_rows
+            }
+          when :size
+            objects.sort! { |obj_a, obj_b|
+              a_size = (obj_a.table.nil? ? 0 : obj_a.table.rows_and_size.fetch(:size)) || 0
+              b_size = (obj_b.table.nil? ? 0 : obj_b.table.rows_and_size.fetch(:size)) || 0
+              b_size <=> a_size
+            }
+        end
+        objects
+      end
+
+      # Note: Not implemented ascending order for now
+      def order_by_related_attribute(dataset, criteria)
+        @can_paginate = false
+        @lazy_order_by = criteria
         dataset
       end
 
+      def order_by_base_attribute(dataset, criteria)
+        @can_paginate = true
+        dataset.order(Sequel.send(:desc, criteria))
+      end
 
-      def order(dataset, criteria={})
+      # Allows to order by any CartoDB::Visualization::Member attribute (eg: updated_at, created_at), plus:
+      # - likes
+      # - mapviews
+      # - row_count
+      # - size
+      def order(dataset, criteria=nil)
         return dataset if criteria.nil? || criteria.empty?
-        dataset.order(*order_params_from(criteria))
+        criteria = criteria.to_sym
+        if ORDERING_RELATED_ATTRIBUTES.include? criteria
+          order_by_related_attribute(dataset, criteria)
+        else
+          order_by_base_attribute(dataset, criteria)
+        end
       end
 
       def filter_by_tags(dataset, tags=[])
@@ -212,10 +384,6 @@ module CartoDB
 
       def tags_from(filters={})
         filters.delete(:tags).to_s.split(',')
-      end
-
-      def order_params_from(criteria)
-        criteria.map { |key, order| Sequel.send(order.to_sym, key.to_sym) }
       end
 
     end