+++ services/datasources/lib/datasources/url/gdrive.rb
@@ -80,7 +80,8 @@ module CartoDB
         # @return string | nil
         def get_auth_url(use_callback_flow=true)
           if use_callback_flow
-            @client.authorization.state = CALLBACK_STATE_DATA_PLACEHOLDER.sub('user', @user.username).sub('service', DATASOURCE_NAME)
+            @client.authorization.state = CALLBACK_STATE_DATA_PLACEHOLDER.sub('user', @user.username)
+                                                                         .sub('service', DATASOURCE_NAME)
           else
             @client.authorization.redirect_uri = REDIRECT_URI
           end